rename persist swap to swap_segwit
