+++ libs/sdk-core/src/persist/mod.rs
@@ -6,7 +6,7 @@ pub(crate) mod migrations;
 pub(crate) mod reverseswap;
 pub(crate) mod send_pays;
 pub(crate) mod settings;
-pub(crate) mod swap;
+pub(crate) mod swap_segwit;
 pub(crate) mod sync;
 pub(crate) mod transactions;
